Restructure wasm-chain-integration to be explicit about version number.
