+++ wasm-chain-integration/src/validation_tests.rs
@@ -10,14 +10,14 @@ use wasm_transform::{
     utils::instantiate,
 };
 
-use crate::ProcessedImports;
+use crate::v0::ProcessedImports;
 
 #[test]
 fn global_offset_test() {
     // This module uses _constant_ global offsets for both data and elem sections.
     let contract = std::fs::read("../testdata/contracts/global-offset-test.wasm").unwrap();
     let res: anyhow::Result<Artifact<ProcessedImports, CompiledFunction>> =
-        instantiate(&crate::ConcordiumAllowedImports, &contract);
+        instantiate(&crate::v0::ConcordiumAllowedImports, &contract);
     assert!(res.is_ok(), "Mutable global offsets allowed in data and elem sections: {:?}", res);
 }
 
@@ -27,7 +27,7 @@ fn mut_global_offset_test() {
     // sections.
     let contract = std::fs::read("../testdata/contracts/mut-global-offset-test.wasm").unwrap();
     let res: anyhow::Result<Artifact<ProcessedImports, CompiledFunction>> =
-        instantiate(&crate::ConcordiumAllowedImports, &contract);
+        instantiate(&crate::v0::ConcordiumAllowedImports, &contract);
 
     assert!(res.is_err(), "Mutable global offsets _not_ allowed in data and elem sections.");
 }
@@ -37,6 +37,6 @@ fn init_global_with_ref_test() {
     // This module tries to instantiate globals using references to other globals.
     let contract = std::fs::read("../testdata/contracts/init-global-with-ref-test.wasm").unwrap();
     let res: anyhow::Result<Artifact<ProcessedImports, CompiledFunction>> =
-        instantiate(&crate::ConcordiumAllowedImports, &contract);
+        instantiate(&crate::v0::ConcordiumAllowedImports, &contract);
     assert!(res.is_err(), "Globals cannot be initialized with references to other globals.");
 }
